Make sure we are using the latest base images
In a CI environment, it is unlikely that we need `docker --pull`.
We probably don't have the base images cached. However, if someone
is using locally then we have a much higher chance.

This commit adds a `docker --pull` for all build command to make
sure we are using the latest base images when we build.