Update SSL images
Previously there was a single "with-ssl" image that we used in shared locations.
However, it gave no indication as to what the ssl was. We now have more than 1
project that needs to test with both libressl and openssl. This renaming moves
us in that direction. Additional changes will be needed but this is a good start.