Create new OpenSSL testing image where we build 1.1.1 from source
This allows us to not have the version change out from underneath
us as well as turn on algos that aren't often on in a particular
distros version.

Note, until such time as it is deprecated, the old "with-openssl-*"
environment will continue to be maintained. Probably for a couple of
weeks as we transition off and let folks who might be using it know.

Closes #9