Rebuild linux builder images after new nightly docker image is created
With this change, we move from rebuilding linux builder images on a set
daily schedule to doing when we get a respository dispatch event from the
ponyc repo that a release has been done.

Currently ponyc will only send an event for a nightly release as such,
this commit breaks updating when there is a new release. Once this runs
for a day and I know it works, I'll be adding in support for also doing
on a regular release, not just a nightly.

This is an intermediate step in getting all this to work as it is the first
time that we are using repository dispatch sending within our own repos.