Temporarily turn off updating some release images
We are in a transition period with the Alpine ponyc image being
based on Alpine 3.18. We can build with nightly tags now but need
to wait until after the next ponyc release in a couple days before
we do release images. We'll turn on release images just before
the next release.